change to float32

+++ rescue_state_bc/rescue_node.py
@@ -5,7 +5,7 @@ from rclpy.lifecycle import TransitionCallbackReturn
 from lifecycle_msgs.msg import Transition
 from enum import Enum, auto
 from robot_msgs.msg import Detections, LEDCommand
-from std_msgs.msg import Int32, Bool, ColorRGBA
+from std_msgs.msg import Int32, Bool, ColorRGBA, Float32
 from sensor_msgs.msg import LaserScan
 from rclpy.action import ActionClient
 from robot_msgs.action import MoveTime
@@ -182,8 +182,8 @@ class Rescue(LifecycleNode):
         self.scan_pub = self.create_publisher(LaserScan, '/scan', 10)
         self.led_cmd_pub = self.create_publisher(LEDCommand, 'led_command', 10)
         self.cmd_vel_pub = self.create_publisher(Twist, '/cmd_vel', 10)
-        self.claw_pub = self.create_publisher(Int32, '/servo/claw', 10) # 0.5 is open, 1 is closed
-        self.lift_pub = self.create_publisher(Int32, '/servo/lift', 10) # up is 2.5, down is 0.2
+        self.claw_pub = self.create_publisher(Float32, '/servo/claw', 10) # 0.5 is open, 1 is closed
+        self.lift_pub = self.create_publisher(Float32, '/servo/lift', 10) # up is 2.5, down is 0.2
 
         return TransitionCallbackReturn.SUCCESS
